refactor(interactions): replace string-discriminated dispatch in selection-handlers with strategy map (#452)
Extract per-tool selection logic into co-located strategy files:
marquee-strategy.ts, wand-strategy.ts, lasso-strategy.ts, and
magnetic-lasso-strategy.ts. Replace 10 if(state.tool === '...')
branches with a STRATEGY_MAP dispatch. selection-handlers.ts drops
from 490 to 177 lines. Adding a new selection tool is now a
one-file-plus-registry change.
